[#7] Remove redundant code
